+++ sdk/resourcemanager/elasticsan/armelasticsan/models.go
@@ -126,55 +126,6 @@ type NetworkRuleSet struct {
 	VirtualNetworkRules []*VirtualNetworkRule
 }
 
-// Operation - Details of a REST API operation, returned from the Resource Provider Operations API
-type Operation struct {
-	// Localized display information for this particular operation.
-	Display *OperationDisplay
-
-	// READ-ONLY; Enum. Indicates the action type. "Internal" refers to actions that are for internal only APIs.
-	ActionType *ActionType
-
-	// READ-ONLY; Whether the operation applies to data-plane. This is "true" for data-plane operations and "false" for ARM/control-plane
-	// operations.
-	IsDataAction *bool
-
-	// READ-ONLY; The name of the operation, as per Resource-Based Access Control (RBAC). Examples: "Microsoft.Compute/virtualMachines/write",
-	// "Microsoft.Compute/virtualMachines/capture/action"
-	Name *string
-
-	// READ-ONLY; The intended executor of the operation; as in Resource Based Access Control (RBAC) and audit logs UX. Default
-	// value is "user,system"
-	Origin *Origin
-}
-
-// OperationDisplay - Localized display information for this particular operation.
-type OperationDisplay struct {
-	// READ-ONLY; The short, localized friendly description of the operation; suitable for tool tips and detailed views.
-	Description *string
-
-	// READ-ONLY; The concise, localized friendly name for the operation; suitable for dropdowns. E.g. "Create or Update Virtual
-	// Machine", "Restart Virtual Machine".
-	Operation *string
-
-	// READ-ONLY; The localized friendly form of the resource provider name, e.g. "Microsoft Monitoring Insights" or "Microsoft
-	// Compute".
-	Provider *string
-
-	// READ-ONLY; The localized friendly name of the resource type related to this operation. E.g. "Virtual Machines" or "Job
-	// Schedule Collections".
-	Resource *string
-}
-
-// OperationListResult - A list of REST API operations supported by an Azure Resource Provider. It contains an URL link to
-// get the next set of results.
-type OperationListResult struct {
-	// READ-ONLY; URL to get the next set of operation list results (if there are any).
-	NextLink *string
-
-	// READ-ONLY; List of operations supported by the resource provider
-	Value []*Operation
-}
-
 // PrivateEndpoint - Response for PrivateEndpoint
 type PrivateEndpoint struct {
 	// READ-ONLY; The ARM identifier for Private Endpoint
@@ -546,6 +497,9 @@ type VolumeGroupProperties struct {
 	// Encryption Properties describing Key Vault and Identity information
 	EncryptionProperties *EncryptionProperties
 
+	// A boolean indicating whether or not Data Integrity Check is enabled
+	EnforceDataIntegrityCheckForIscsi *bool
+
 	// A collection of rules governing the accessibility from specific network locations.
 	NetworkACLs *NetworkRuleSet
 
@@ -575,6 +529,9 @@ type VolumeGroupUpdateProperties struct {
 
 	// Encryption Properties describing Key Vault and Identity information
 	EncryptionProperties *EncryptionProperties
+
+	// A boolean indicating whether or not Data Integrity Check is enabled
+	EnforceDataIntegrityCheckForIscsi *bool
 
 	// A collection of rules governing the accessibility from specific network locations.
 	NetworkACLs *NetworkRuleSet