+++ sdk/resourcemanager/elasticsan/armelasticsan/responses.go
@@ -43,12 +43,6 @@ type ElasticSansClientUpdateResponse struct {
 	ElasticSan
 }
 
-// OperationsClientListResponse contains the response from method OperationsClient.NewListPager.
-type OperationsClientListResponse struct {
-	// A list of REST API operations supported by an Azure Resource Provider. It contains an URL link to get the next set of results.
-	OperationListResult
-}
-
 // PrivateEndpointConnectionsClientCreateResponse contains the response from method PrivateEndpointConnectionsClient.BeginCreate.
 type PrivateEndpointConnectionsClientCreateResponse struct {
 	// Response for PrivateEndpoint Connection object